+++ AL-Game/src/com/aionemu/gameserver/model/house/HouseBidEntry.java
@@ -1,29 +1,26 @@
-/**
- * This file is part of Aion-Lightning <aion-lightning.org>.
+/*
+ * This file is part of Encom. **ENCOM FUCK OTHER SVN**
  *
- *  Aion-Lightning is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU General Public License as published by
+ *  Encom is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU Lesser Public License as published by
  *  the Free Software Foundation, either version 3 of the License, or
  *  (at your option) any later version.
  *
- *  Aion-Lightning is distributed in the hope that it will be useful,
+ *  Encom is distributed in the hope that it will be useful,
  *  but WITHOUT ANY WARRANTY; without even the implied warranty of
  *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU General Public License for more details. *
- *  You should have received a copy of the GNU General Public License
- *  along with Aion-Lightning.
- *  If not, see <http://www.gnu.org/licenses/>.
+ *  GNU Lesser Public License for more details.
+ *
+ *  You should have received a copy of the GNU Lesser Public License
+ *  along with Encom.  If not, see <http://www.gnu.org/licenses/>.
  */
 package com.aionemu.gameserver.model.house;
 
 import com.aionemu.gameserver.configs.main.HousingConfig;
 import com.aionemu.gameserver.model.templates.housing.HouseType;
 
-/**
- * @author Rolandas
- */
-public class HouseBidEntry implements Cloneable {
-
+public class HouseBidEntry implements Cloneable
+{
 	private int entryIndex;
 	private int landId;
 	private int address;
@@ -35,7 +32,7 @@ public class HouseBidEntry implements Cloneable {
 	private int mapId;
 	private int lastBiddingPlayer;
 	private long lastBidTime;
-
+	
 	public HouseBidEntry(House house, int index, long initialBid) {
 		entryIndex = index;
 		landId = house.getLand().getId();
@@ -47,82 +44,82 @@ public class HouseBidEntry implements Cloneable {
 		lastBiddingPlayer = 0;
 		lastBidTime = 0;
 	}
-
+	
 	private HouseBidEntry() {
 	}
-
+	
 	public int getEntryIndex() {
 		return entryIndex;
 	}
-
+	
 	public void setEntryIndex(int entryIndex) {
 		this.entryIndex = entryIndex;
 	}
-
+	
 	public int getLandId() {
 		return landId;
 	}
-
+	
 	public int getAddress() {
 		return address;
 	}
-
+	
 	public int getBuildingId() {
 		return buildingId;
 	}
-
+	
 	public void setBuildingId(int buildingId) {
 		this.buildingId = buildingId;
 	}
-
+	
 	public long getBidPrice() {
 		return bidPrice;
 	}
-
+	
 	public void setBidPrice(long bidPrice) {
 		this.bidPrice = bidPrice;
 	}
-
+	
 	public int getBidCount() {
 		return bidCount;
 	}
-
+	
 	public void incrementBidCount() {
 		this.bidCount++;
 	}
-
+	
 	public final long getUnk2() {
 		return unk2;
 	}
-
+	
 	public HouseType getHouseType() {
 		return houseType;
 	}
-
+	
 	public int getMapId() {
 		return mapId;
 	}
-
+	
 	public int getLastBiddingPlayer() {
 		return lastBiddingPlayer;
 	}
-
+	
 	public void setLastBiddingPlayer(int lastBiddingPlayer) {
 		this.lastBiddingPlayer = lastBiddingPlayer;
 	}
-
+	
 	public long getRefundKinah() {
-		return (long) (bidPrice * HousingConfig.BID_REFUND_PERCENT);
+		return (long) (bidPrice * (float)HousingConfig.BID_REFUND_PERCENT);
 	}
-
+	
 	public long getLastBidTime() {
 		return lastBidTime;
 	}
-
+	
 	public void setLastBidTime(long lastBidTime) {
 		this.lastBidTime = lastBidTime;
 	}
-
+	
 	public Object Clone() {
 		HouseBidEntry cloned = new HouseBidEntry();
 		cloned.address = this.address;
