+++ AL-Game/src/com/aionemu/gameserver/network/aion/clientpackets/CM_CLOSE_DIALOG.java
@@ -1,22 +1,21 @@
-/**
- * This file is part of Aion-Lightning <aion-lightning.org>.
+/*
+ * This file is part of aion-unique <aionunique.smfnew.com>.
  *
- *  Aion-Lightning is free software: you can redistribute it and/or modify
+ *  aion-unique is free software: you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
  *  the Free Software Foundation, either version 3 of the License, or
  *  (at your option) any later version.
  *
- *  Aion-Lightning is distributed in the hope that it will be useful,
+ *  aion-unique is distributed in the hope that it will be useful,
  *  but WITHOUT ANY WARRANTY; without even the implied warranty of
  *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU General Public License for more details. *
+ *  GNU General Public License for more details.
+ *
  *  You should have received a copy of the GNU General Public License
- *  along with Aion-Lightning.
- *  If not, see <http://www.gnu.org/licenses/>.
+ *  along with aion-unique.  If not, see <http://www.gnu.org/licenses/>.
  */
 package com.aionemu.gameserver.network.aion.clientpackets;
 
-import com.aionemu.commons.network.util.ThreadPoolManager;
 import com.aionemu.gameserver.ai2.event.AIEventType;
 import com.aionemu.gameserver.model.gameobjects.Npc;
 import com.aionemu.gameserver.model.gameobjects.VisibleObject;
@@ -27,59 +26,39 @@ import com.aionemu.gameserver.network.aion.AionConnection.State;
 import com.aionemu.gameserver.network.aion.serverpackets.SM_HEADING_UPDATE;
 import com.aionemu.gameserver.services.DialogService;
 import com.aionemu.gameserver.services.player.PlayerMailboxState;
+import com.aionemu.gameserver.utils.ThreadPoolManager;
 
-public class CM_CLOSE_DIALOG extends AionClientPacket {
-
-	/**
-	 * Target object id that client wants to TALK WITH or 0 if wants to unselect
-	 */
+public class CM_CLOSE_DIALOG extends AionClientPacket
+{
 	private int targetObjectId;
-
-	/**
-	 * Constructs new instance of <tt>CM_CM_REQUEST_DIALOG </tt> packet
-	 *
-	 * @param opcode
-	 */
+	
 	public CM_CLOSE_DIALOG(int opcode, State state, State... restStates) {
 		super(opcode, state, restStates);
 	}
-
-	/**
-	 * {@inheritDoc}
-	 */
+	
 	@Override
 	protected void readImpl() {
 		targetObjectId = readD();
 	}
-
-	/**
-	 * {@inheritDoc}
-	 */
+	
 	@Override
 	protected void runImpl() {
 		Player player = getConnection().getActivePlayer();
 		final VisibleObject obj = player.getKnownList().getObject(targetObjectId);
 		final AionConnection client = getConnection();
-		if (obj == null) {
+		if(obj == null) {
 			return;
-		}
-
-		if (obj instanceof Npc) {
-			Npc npc = (Npc) obj;
+		} if (obj instanceof Npc) {
+			Npc npc = (Npc)obj;
 			npc.getAi2().onCreatureEvent(AIEventType.DIALOG_FINISH, player);
 			DialogService.onCloseDialog(npc, player);
-
 			ThreadPoolManager.getInstance().schedule(new Runnable() {
-
 				@Override
 				public void run() {
-					client.sendPacket(new SM_HEADING_UPDATE(targetObjectId, obj.getHeading()));
+					client.sendPacket(new SM_HEADING_UPDATE(targetObjectId, (byte)obj.getHeading()));
 				}
 			}, 1200);
-
-		}
-
-		if (player.getMailbox().mailBoxState != 0) {
+		} if (player.getMailbox().mailBoxState != 0) {
 			player.getMailbox().mailBoxState = PlayerMailboxState.CLOSED;
 		}
 	}
