+++ AL-Game/src/com/aionemu/gameserver/utils/SafeMath.java
@@ -1,18 +1,18 @@
-/**
- * This file is part of Aion-Lightning <aion-lightning.org>.
+/*
+ * This file is part of aion-lightning <aion-lightning.com>.
  *
- *  Aion-Lightning is free software: you can redistribute it and/or modify
+ *  aion-lightning is free software: you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
  *  the Free Software Foundation, either version 3 of the License, or
  *  (at your option) any later version.
  *
- *  Aion-Lightning is distributed in the hope that it will be useful,
+ *  aion-lightning is distributed in the hope that it will be useful,
  *  but WITHOUT ANY WARRANTY; without even the implied warranty of
  *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU General Public License for more details. *
+ *  GNU General Public License for more details.
+ *
  *  You should have received a copy of the GNU General Public License
- *  along with Aion-Lightning.
- *  If not, see <http://www.gnu.org/licenses/>.
+ *  along with aion-lightning.  If not, see <http://www.gnu.org/licenses/>.
  */
 package com.aionemu.gameserver.utils;
 
@@ -22,26 +22,28 @@ package com.aionemu.gameserver.utils;
 public class SafeMath {
 
 	public static int addSafe(int source, int value) throws OverfowException {
-		long s = (long) source + (long) value;
-		if (s < Integer.MIN_VALUE || s > Integer.MAX_VALUE) {
-			throw new OverfowException(source + " + " + value + " = " + ((long) source + (long) value));
-		}
-		return (int) s;
+		 long s = (long)source+(long)value;
+     if (s < Integer.MIN_VALUE ||
+         s > Integer.MAX_VALUE) {
+         throw new OverfowException(source + " + " + value + " = " + ((long) source + (long) value));
+     }
+     return (int)s;
 	}
 
 	public static long addSafe(long source, long value) throws OverfowException {
 		if ((source > 0 && value > Long.MAX_VALUE - source) || (source < 0 && value < Long.MIN_VALUE - source)) {
-			throw new OverfowException(source + " + " + value + " = " + (source + value));
+			throw new OverfowException(source + " + " + value + " = " + ((long) source + (long) value));
 		}
 		return source + value;
 	}
 
 	public static int multSafe(int source, int value) throws OverfowException {
-		long m = ((long) source) * ((long) value);
-		if (m < Integer.MIN_VALUE || m > Integer.MAX_VALUE) {
-			throw new OverfowException(source + " * " + value + " = " + ((long) source * (long) value));
-		}
-		return (int) m;
+		 long m = ((long)source)*((long)value);
+     if (m < Integer.MIN_VALUE ||
+         m > Integer.MAX_VALUE) {
+         throw new OverfowException(source + " * " + value + " = " + ((long) source * (long) value));
+     }
+     return (int)m;
 	}
 
 	public static long multSafe(long a, long b) throws OverfowException {
